Make `gn_version` available in the dotfile.
It would be useful if `gn_version` was available in the dotfile
so that you can check for a specific version there before anything
else happens (the dotfile is processed before args.gn, BUILDCONFIG,
or any BUILD.gn file).

This CL restructures things to make that happen.

Note that it appears that this is the only variable that is really
safe to be defined in the dotfile and args.gn scopes, unfortunately.
It seems like it would be useful if at least `host_os` and `host_cpu`
were also available, but since those values can be modified via
`default_args` in the dotfile, it would be hard to reason about the
values for them.

Change-Id: Ia3c98e0288829e9f618e82591506ca805684618c
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/18521
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: Takuto Ikuta <tikuta@google.com>
Commit-Queue: David Turner <digit@google.com>